+++ front-accessibility/scripts/_argparse.py
@@ -24,7 +24,7 @@ import argparse
 from typing import Optional
 
 
-SKILL_VERSION = "0.22.0"
+SKILL_VERSION = "0.23.0"
 
 
 def make_parser(